修改编辑器设置，削减不必要的警告和提示数量。 新建测试项目，准备尝试修复ListView控件

+++ src/Wpf.Ui.Demo.Mvvm/ViewModels/DataViewModel.cs
@@ -35,7 +35,7 @@ public partial class DataViewModel : ObservableObject, INavigationAware
                 {
                     Color = new SolidColorBrush(
                         Color.FromArgb(
-                            (byte)200,
+							200,
                             (byte)random.Next(0, 250),
                             (byte)random.Next(0, 250),
                             (byte)random.Next(0, 250)
